refactor(runtime): extract tool-result-archive and shared source-ref contract
Move the archived-tool-result archive subsystem and the shared source-ref
contract out of synthesis-cache.ts, keeping behavior neutral:

- extract the archive lifecycle (placeholder/candidate/ref schema,
  serialization, retrieve/prune/collect, retrieval result) to
  tool-result-archive.ts; archive retrieval now returns the narrow
  ArchivedToolResultSourceRef[] instead of the full SynthesisSourceRef union
- extract the 4-kind ContextSourceRef contract (+ SynthesisSourceRef alias +
  isValidSynthesisSourceRef) to context-source-ref.ts, since it is produced by
  archive, history-compact, and history-search, not synthesis alone
- relocate ActiveToolResultPrunePolicy/ActiveToolResultArchiveCandidate/
  ActiveArchivedToolResultPlaceholder/ACTIVE_ARCHIVED_TOOL_RESULT_PLACEHOLDER_KIND/
  ActiveArchivedToolResultReason to active-tool-result-prune.ts (where the
  implementation already lives)
- relocate HistoryCompactSourceArchiveRef + runtimeEventArchiveBody/Sha256 +
  isValidHistoryCompactSourceArchiveRef to history-compact.ts (sole consumer)
- relocate the cross-block pure helpers utf8ByteLength and
  optionalNonNegativeFiniteNumber to context-budget-helpers.ts
- fix pre-existing reverse imports: active-tool-result-prune, active-full-
  compact, history-compact-artifacts, history-compact-cleanup now import their
  owners directly instead of the context-budget compatibility layer

No reverse imports: tool-result-archive and context-source-ref only import the
leaf helpers and @maka/core; history-compact no longer imports synthesis-cache;
synthesis-cache imports the archive and source-ref leaves forward. context-
budget re-exports every moved public symbol for import compatibility.

Part of #1084.
