+++ packages/runtime/src/history-compact-cleanup.ts
@@ -2,12 +2,12 @@ import { Buffer } from 'node:buffer';
 import { createHash } from 'node:crypto';
 import type { ArtifactRecord } from '@maka/core';
 import type { RuntimeEvent } from '@maka/core/runtime-event';
+import { estimateRuntimeEventsTokens } from './context-budget-helpers.js';
 import {
-  estimateRuntimeEventsTokens,
   validateHistoryCompactBlockShape,
   type HistoryCompactBlock,
   type HistoryCompactSourceArchiveRef,
-} from './context-budget.js';
+} from './history-compact.js';
 import type { HistoryCompactArtifactStore } from './history-compact-artifacts.js';
 import {
   matchHistoryCompactCheckpointPrefix,
